requirements.txt
Django>=3.0.10,<=4
djangorestframework>=3.11.1
psycopg2-binary>=2.8